+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - The `--llm-name` (and `--judge-model-name`) argument can now also be a module path like `eval_framework.llm.huggingface.HFLLM`.
   Combining this with `--llm-args` (`-judge-model-args`) should cover many use-cases without having to provide a `models.py` file.
+- Minor fix allowing for empty `stop_sequences` in `eval_framework.llm.huggingface.StopSequenceCriteria`.
 
 ### Tasks